# specs/002-firmware-catalog/contracts/rest-openapi.yaml
openapi: 3.1.0
info:
  title: GARD — Firmware Catalog (F2 additions to the REST surface)
  version: 0.2.0
  description: |
    F2 augments the F1 REST API with firmware-catalog read endpoints, a single
    per-device compliance read, and an optional blob upload/download pair.
    All `/api/v1/firmware/...` and `/api/v1/devices/{id}/firmware-compliance`
    endpoints require Bearer auth with the `READ_FIRMWARE_CATALOG` permission
    (or higher); the blob upload requires `MANAGE_FIRMWARE_BLOB`. There is no
    `MANAGE_FIRMWARE_CATALOG` permission — catalog mutation is by git push,
    not by role.
servers:
  - url: http://127.0.0.1:8080
    description: Local dev (compose stack)
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            details: {type: object, nullable: true, additionalProperties: true}
            correlation_id: {type: string, nullable: true}

    # ----------- Scope selector (shared with catalog YAML) -----------
    ScopeSelector:
      type: object
      minProperties: 1
      properties:
        vendor_normalized: {type: string}
        platform_family: {type: string}
        region_in: {type: array, items: {type: string}, minItems: 1}
        site_in: {type: array, items: {type: string}, minItems: 1}
        role_in: {type: array, items: {type: string}, minItems: 1}
        hardware_revision_in: {type: array, items: {type: string}, minItems: 1}
        not_in_state:
          type: array
          items:
            type: string
            enum: [imported, classified, target_defined, compliant, outside_target, manual_review_required]
          minItems: 1
        tagged_with: {type: array, items: {type: string}, minItems: 1}
      additionalProperties: false

    # ----------- Firmware target -----------
    FirmwareTarget:
      type: object
      required: [id, name, platform_family, target_version, scope_selector, loaded_at, catalog_schema_version]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        platform_family: {type: string}
        target_version: {type: string}
        scope_selector: {$ref: "#/components/schemas/ScopeSelector"}
        valid_from: {type: string, format: date-time, nullable: true}
        valid_until: {type: string, format: date-time, nullable: true}
        notes: {type: string, nullable: true}
        loaded_at: {type: string, format: date-time}
        loaded_from_git_sha: {type: string, nullable: true}
        source_file_relpath: {type: string}
        catalog_schema_version: {type: string}
    FirmwareTargetList:
      type: object
      required: [items, total_returned]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/FirmwareTarget"}}
        total_returned: {type: integer, minimum: 0}
        next_page_token: {type: string, nullable: true}

    # ----------- Firmware package -----------
...
security:
  - HTTPBearer: []
paths:
...
  /api/v1/firmware/targets:
    get:
      tags: [firmware]
      summary: List firmware targets
      operationId: listFirmwareTargets
      parameters:
        - in: query
          name: platform_family
          schema: {type: string}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 500, default: 50}
        - in: query
          name: page_token
          schema: {type: string, nullable: true}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FirmwareTargetList"}
        "401":
          description: Missing or invalid bearer
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
        "403":
          description: Missing READ_FIRMWARE_CATALOG
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
  /api/v1/firmware/targets/{id}:
    get:
      tags: [firmware]
      summary: Get one firmware target
      operationId: getFirmwareTarget
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/FirmwareTarget"}}}
        "404":
          description: Not found
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}

  # =========================== firmware/packages ===========================
...